pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apex-predict"
version = "0.1.0"
description = "Apex Predict backend + worker + mobile skeleton"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.37",
  "alembic>=1.16.0",
  "asyncpg>=0.30.0",
  "aiosqlite>=0.20.0",
  "httpx>=0.28.0",
  "pyjwt[crypto]>=2.10.1",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.0",
  "mypy>=1.14.0",
  "hypothesis>=6.125.0",
]
load = [
  "locust>=2.32.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["apex_predict*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-q"
markers = [
  "unit: fast deterministic tests for isolated functions",
  "security: authentication/authorization/input-safety tests",
  "penetration: attack-simulation tests for abuse and bypass attempts",
  "fuzz: randomized malformed-input tests",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["UP017", "UP042"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
